refactor: remove max-lines baseline exceptions
