+++ packages/gateway/src/modules/node/capability-catalog-browser.ts
@@ -1,0 +1,186 @@
+import {
+  BrowserClickArgs,
+  BrowserClickResult,
+  BrowserCloseArgs,
+  BrowserCloseResult,
+  BrowserConsoleMessagesArgs,
+  BrowserConsoleMessagesResult,
+  BrowserDragArgs,
+  BrowserDragResult,
+  BrowserEvaluateArgs,
+  BrowserEvaluateResult,
+  BrowserFillFormArgs,
+  BrowserFillFormResult,
+  BrowserHandleDialogArgs,
+  BrowserHandleDialogResult,
+  BrowserHoverArgs,
+  BrowserHoverResult,
+  BrowserLaunchArgs,
+  BrowserLaunchResult,
+  BrowserNavigateArgs,
+  BrowserNavigateBackArgs,
+  BrowserNavigateBackResult,
+  BrowserNavigateResult,
+  BrowserNetworkRequestsArgs,
+  BrowserNetworkRequestsResult,
+  BrowserPressKeyArgs,
+  BrowserPressKeyResult,
+  BrowserResizeArgs,
+  BrowserResizeResult,
+  BrowserRunCodeArgs,
+  BrowserRunCodeResult,
+  BrowserScreenshotArgs,
+  BrowserScreenshotResult,
+  BrowserSelectOptionArgs,
+  BrowserSelectOptionResult,
+  BrowserSnapshotArgs,
+  BrowserSnapshotResult,
+  BrowserTabsArgs,
+  BrowserTabsResult,
+  BrowserTypeArgs,
+  BrowserTypeResult,
+  BrowserUploadFileArgs,
+  BrowserUploadFileResult,
+  BrowserWaitForArgs,
+  BrowserWaitForResult,
+} from "@tyrum/contracts";
+import { ba, type CapabilityCatalogEntry } from "./capability-catalog-helpers.js";
+
+export const BROWSER_CAPABILITY_CATALOG_ENTRIES: readonly CapabilityCatalogEntry[] = [
+  ba(
+    "tyrum.browser.launch",
+    "launch",
+    "Launch a browser session.",
+    BrowserLaunchArgs,
+    BrowserLaunchResult,
+  ),
+  ba(
+    "tyrum.browser.navigate",
+    "navigate",
+    "Navigate to a URL.",
+    BrowserNavigateArgs,
+    BrowserNavigateResult,
+  ),
+  ba(
+    "tyrum.browser.navigate-back",
+    "navigate_back",
+    "Navigate back in history.",
+    BrowserNavigateBackArgs,
+    BrowserNavigateBackResult,
+  ),
+  ba(
+    "tyrum.browser.snapshot",
+    "snapshot",
+    "Collect a page accessibility snapshot.",
+    BrowserSnapshotArgs,
+    BrowserSnapshotResult,
+  ),
+  ba("tyrum.browser.click", "click", "Click a page element.", BrowserClickArgs, BrowserClickResult),
+  ba(
+    "tyrum.browser.type",
+    "type",
+    "Type text into an element.",
+    BrowserTypeArgs,
+    BrowserTypeResult,
+  ),
+  ba(
+    "tyrum.browser.fill-form",
+    "fill_form",
+    "Fill a form field.",
+    BrowserFillFormArgs,
+    BrowserFillFormResult,
+  ),
+  ba(
+    "tyrum.browser.select-option",
+    "select_option",
+    "Select from a dropdown.",
+    BrowserSelectOptionArgs,
+    BrowserSelectOptionResult,
+  ),
+  ba(
+    "tyrum.browser.hover",
+    "hover",
+    "Hover over an element.",
+    BrowserHoverArgs,
+    BrowserHoverResult,
+  ),
+  ba("tyrum.browser.drag", "drag", "Drag an element.", BrowserDragArgs, BrowserDragResult),
+  ba(
+    "tyrum.browser.press-key",
+    "press_key",
+    "Press a keyboard key.",
+    BrowserPressKeyArgs,
+    BrowserPressKeyResult,
+  ),
+  ba(
+    "tyrum.browser.screenshot",
+    "screenshot",
+    "Capture a page screenshot.",
+    BrowserScreenshotArgs,
+    BrowserScreenshotResult,
+  ),
+  ba(
+    "tyrum.browser.evaluate",
+    "evaluate",
+    "Run JavaScript in page context.",
+    BrowserEvaluateArgs,
+    BrowserEvaluateResult,
+  ),
+  ba(
+    "tyrum.browser.wait-for",
+    "wait_for",
+    "Wait for a page condition.",
+    BrowserWaitForArgs,
+    BrowserWaitForResult,
+  ),
+  ba(
+    "tyrum.browser.tabs",
+    "tabs",
+    "List or switch browser tabs.",
+    BrowserTabsArgs,
+    BrowserTabsResult,
+  ),
+  ba(
+    "tyrum.browser.upload-file",
+    "upload_file",
+    "Upload a file to a file input.",
+    BrowserUploadFileArgs,
+    BrowserUploadFileResult,
+  ),
+  ba(
+    "tyrum.browser.console-messages",
+    "console_messages",
+    "Read browser console output.",
+    BrowserConsoleMessagesArgs,
+    BrowserConsoleMessagesResult,
+  ),
+  ba(
+    "tyrum.browser.network-requests",
+    "network_requests",
+    "Inspect network requests.",
+    BrowserNetworkRequestsArgs,
+    BrowserNetworkRequestsResult,
+  ),
+  ba(
+    "tyrum.browser.resize",
+    "resize",
+    "Resize the browser viewport.",
+    BrowserResizeArgs,
+    BrowserResizeResult,
+  ),
+  ba("tyrum.browser.close", "close", "Close the browser.", BrowserCloseArgs, BrowserCloseResult),
+  ba(
+    "tyrum.browser.handle-dialog",
+    "handle_dialog",
+    "Accept or dismiss a dialog.",
+    BrowserHandleDialogArgs,
+    BrowserHandleDialogResult,
+  ),
+  ba(
+    "tyrum.browser.run-code",
+    "run_code",
+    "Run arbitrary code in the browser.",
+    BrowserRunCodeArgs,
+    BrowserRunCodeResult,
+  ),
+];
